refactor: setModalOpacity props드릴링
